abstract components and clean up old code
