--- client/src/logic/tactical-map.ts
@@ -1,3 +1,0 @@
-import { TacticalMap } from '@cohdex/tactical-map'
-
-export const TMap = new TacticalMap()
