+++ client/src/provider/Container.ts
@@ -1,4 +1,3 @@
-import { TMap } from '../logic/tactical-map'
 import { StrategyService } from '../services/StrategyService'
 
-export const strategyService = new StrategyService(TMap)
+export const strategyService = new StrategyService()
