+++ client/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './preloadImages'
+export * from './replaceTgaWithPng'
