+++ client/src/utils/replaceTgaWithPng.ts
@@ -1,0 +1,3 @@
+export function replaceTgaWithPng(url: string) {
+  return url.replace('tga', 'png')
+}
